// guara/src/types/apiTypes.ts
export interface SparqlBindingValue {
  type: 'uri' | 'literal' | 'bnode';
  value: string;
  datatype?: string; // Ex: "http://www.w3.org/2001/XMLSchema#integer"
  'xml:lang'?: string; // Ex: "pt-BR"
}

// Representa uma linha de resultado, mapeando nomes de variáveis para seus valores
export type SparqlResult = Record<string, SparqlBindingValue>;

// Estrutura completa de uma resposta de consulta SPARQL
export interface SparqlQueryResponse {
  head: {
    vars: string[];
  };
  results: {
    bindings: SparqlResult[];
  };
}
// ...
export interface Relation {
  id?: { value: string };
  property: string;
  value: string;
  valueType: 'uri' | 'literal';
  valueTitle?: string;
}

export interface AddRelationPayload {
  sujeito_id_local: string;
  predicado_uri: string;
  valor_objeto: string;
  tipo_valor_objeto: 'URI' | 'Literal';
  repository_update_url: string;
  repository_base_uri: string;
  literal_datatype_uri?: string;
  literal_lang_tag?: string;
}

export interface RemoveSpecificRelationPayload {
  s: string;
  p: string;
  o: string;
  repository_update_url: string;
}

export interface UpdateObjectPropertiesPayload {
  object_uri_to_update: string;
  titulo: string;
  resumo: string;
  repository_update_url: string;
  descricao?: string;
}